create new look of the base map, make it smalle for better performance of the game and for optimazation and simpliest of the game.
