+++ Data/data.json
@@ -1,60 +1,20 @@
 {
-    "Map1": [
-        {"name": "book_shelf", "type": "other", "rect": "(450, 190, 70, 90)"}, 
-        {"name": "book_shelf", "type": "other", "rect": "(520, 190, 70, 90)"}, 
-        {"name": "monitor_1", "type": "other", "rect": "(650, 180, 50, 50)"}, 
-        {"name": "table_1", "type": "other", "rect": "(515, 400, 50, 50)"}, 
-        {"name": "table_1", "type": "other", "rect": "(1180, 400, 50, 50)"}, 
+    "Base": [
+        {"name": "book_shelf", "type": "other", "rect": "(600, 30, 70, 90)", "comment": "top side"},
+        {"name": "book_shelf", "type": "other", "rect": "(535, 30, 70, 90)"},
+        {"name": "book_shelf", "type": "other", "rect": "(470, 30, 70, 90)"},
+        {"name": "monitor_1", "type": "other", "rect": "(400, 45, 70, 70)"},
+        {"name": "box_1", "type": "other", "rect": "(300, 60, 60, 55)"},
+        {"name": "box_1", "type": "other", "rect": "(240, 60, 60, 55)"},
+        {"name": "table_1", "type": "other", "rect": "(600, 150, 60, 60)"},
 
-        {"name": "monitor_1", "type": "animated", "rect": "(720, 175, 70, 70)", "comment": "animated object"},
-        {"name": "monitor_1", "type": "animated", "rect": "(795, 175, 70, 70)"},
-        {"name": "monitor_1", "type": "animated", "rect": "(870, 175, 70, 70)"},
+        {"name": "monitor_1", "type": "animated", "rect": "(150, 25, 70, 70)"},
+        {"name": "monitor_1", "type": "animated", "rect": "(70, 25, 70, 70)"},
 
-        {"name": "door_1", "type": "animated_once", "rect": "(1330, -10, 90, 90)", "comment": "animated once objects"},
-
-        {"name": "hidden", "type": "hidden", "rect": "(0, 25, 250, 50)"},
-        {"name": "hidden", "type": "hidden", "rect": "(250, 25, 1000, 50)"},
-        {"name": "hidden", "type": "hidden2", "rect": "(225, 25, 25, 300)"},
-        {"name": "hidden", "type": "hidden", "rect": "(225, 300, 25, 70)"},
-        {"name": "hidden", "type": "hidden2", "rect": "(225, 425, 25, 425)"},
-        {"name": "hidden", "type": "hidden2", "rect": "(0, 500, 250, 25)"},
-        {"name": "hidden", "type": "hidden", "rect": "(0, 550, 250, 50)"},
-        {"name": "hidden", "type": "hidden", "rect": "(225, 825, 75, 100)"},
-        {"name": "hidden", "type": "hidden", "rect": "(375, 825, 50, 100)"},
-        {"name": "hidden", "type": "hidden2", "rect": "(0, 1050, 1600, 25)"},
-        {"name": "hidden", "type": "hidden2", "rect": "(425, 675, 25, 400)"},
-        {"name": "hidden", "type": "hidden2", "rect": "(425, 400, 25, 125)"},
-        {"name": "hidden", "type": "hidden2", "rect": "(425, 150, 25, 150)"},
-        {"name": "hidden", "type": "hidden", "rect": "(425, 275, 25, 75)"},
-        {"name": "hidden", "type": "hidden", "rect": "(425, 550, 900, 50)"},
-        {"name": "hidden", "type": "hidden2", "rect": "(425, 500, 900, 25)"},
-        {"name": "hidden", "type": "hidden2", "rect": "(425, 150, 725, 25)"},
-        {"name": "hidden", "type": "hidden2", "rect": "(1200, 150, 125, 25)"},
-        {"name": "hidden", "type": "hidden2", "rect": "(625, 150, 25, 75)"},
-        {"name": "hidden", "type": "hidden", "rect": "(625, 225, 25, 75)"},
-        {"name": "hidden", "type": "hidden", "rect": "(450, 200, 175, 50)"},
-        {"name": "hidden", "type": "hidden2", "rect": "(625, 325, 25, 175)"},
-        {"name": "hidden", "type": "hidden2", "rect": "(1075, 175, 25, 325)"},
-        {"name": "hidden", "type": "hidden", "rect": "(630, 200, 520, 50)"},
-        {"name": "hidden", "type": "hidden2", "rect": "(1300, 175, 25, 325)"},
-        {"name": "hidden", "type": "hidden", "rect": "(1200, 200, 100, 50)"},
-        {"name": "hidden", "type": "hidden2", "rect": "(1325, 225, 100, 25)"},
-        {"name": "hidden", "type": "hidden", "rect": "(1325, 275, 100, 50)"},
-        {"name": "hidden", "type": "hidden", "rect": "(1275, 25, 200, 50)"},
-        {"name": "hidden", "type": "hidden2", "rect": "(1250, 25, 25, 125)"},
-        {"name": "hidden", "type": "hidden2", "rect": "(1475, 25, 25, 1025)"},
-        {"name": "hidden", "type": "hidden2", "rect": "(675, 675, 25, 250)"},
-        {"name": "hidden", "type": "hidden2", "rect": "(1175, 675, 25, 250)"},
-        {"name": "hidden", "type": "hidden2", "rect": "(1325, 675, 25, 375)"},
-        {"name": "hidden", "type": "hidden2", "rect": "(450, 675, 50, 25)", "comment": "for extended door"},
-        {"name": "hidden", "type": "hidden2", "rect": "(575, 675, 300, 25)"},
-        {"name": "hidden", "type": "hidden2", "rect": "(950, 675, 375, 25)"},
-        {"name": "hidden", "type": "hidden2", "rect": "(675, 900, 500, 25)"},
-        {"name": "hidden", "type": "hidden", "rect": "(675, 950, 525, 50)"},
-        {"name": "hidden", "type": "hidden", "rect": "(450, 725, 50, 50)", "comment": "for extended door"},
-        {"name": "hidden", "type": "hidden", "rect": "(575, 725, 100, 50)"},
-        {"name": "hidden", "type": "hidden", "rect": "(700, 725, 175, 50)"},
-        {"name": "hidden", "type": "hidden", "rect": "(950, 725, 400, 50)"}
+        {"name": "hidden", "type": "hidden2", "rect": "(0, 0, 25, 400)", "comment": "left side"},
+        {"name": "hidden", "type": "hidden2", "rect": "(675, 0, 25, 175)", "comment": "right side"},
+        {"name": "hidden", "type": "hidden2", "rect": "(750, 150, 25, 175)", "comment": "door 2"},
+        {"name": "hidden", "type": "hidden2", "rect": "(675, 300, 75, 25)", "comment": "door 2"},
+        {"name": "hidden", "type": "hidden2", "rect": "(675, 175, 75, 25)", "comment": "door 2"}
     ]
-
 }
