--- lab.py
@@ -1,76 +1,0 @@
-import pygame
-
-class TileMap():
-    def __init__(self, tilesize, x, y):
-        pygame.init()
-        self.tilesize = tilesize
-        self.images = []
-        self.blocs = [] # store all blocs that player can collide
-        self.x = x
-        self.y = y
-        self.name = 'map'
-        self.images.append(pygame.transform.scale(pygame.image.load("C:/Users/Admin/OneDrive/Documents/pygame_project/characters/objects/floor_2.png"), (self.tilesize, self.tilesize)))
-        self.images.append(pygame.transform.scale(pygame.image.load("C:/Users/Admin/OneDrive/Documents/pygame_project/characters/objects/border_3.png"), (self.tilesize, self.tilesize)))
-        self.images.append(pygame.transform.scale(pygame.image.load("C:/Users/Admin/OneDrive/Documents/pygame_project/characters/objects/floor_1.png"), (self.tilesize, self.tilesize)))
-        self.images.append(pygame.transform.scale(pygame.image.load("C:/Users/Admin/OneDrive/Documents/pygame_project/characters/objects/wall_1.png"), (self.tilesize, self.tilesize)))
-        self.images.append(pygame.transform.scale(pygame.image.load("C:/Users/Admin/OneDrive/Documents/pygame_project/characters/objects/border_1.png"), (self.tilesize, self.tilesize)))
-        self.array = [[1,1,1,1,1,1,1,1,1,2,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,2,1,1,1,1,1,1,1,1,2,3,3,3],
-[1,1,1,1,1,1,1,1,1,2,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,2,1,1,1,1,1,1,1,1,2,3,3,3],
-[1,1,1,1,1,1,1,1,1,2,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,2,1,1,1,1,1,1,1,1,2,3,3,3],
-[4,4,4,4,4,4,4,4,4,2,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,2,3,3,3,3,3,3,3,3,2,3,3,3],
-[4,4,4,4,4,4,4,4,4,2,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,2,3,3,3,3,3,3,3,3,2,3,3,3],
-[4,4,4,4,4,4,4,4,4,2,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,2,3,3,3,3,3,3,3,3,2,3,3,3],
-[4,4,4,4,4,4,4,4,4,2,3,3,3,3,3,3,3,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,3,3,5,5,2,5,5,3,3,3,3,3,3,2,3,3,3],
-[4,4,4,4,4,4,4,4,4,2,3,3,3,3,3,3,3,2,1,1,1,1,1,1,1,2,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,2,1,1,3,3,1,1,1,1,2,3,3,3,3,3,3,2,3,3,3],
-[4,4,4,4,4,4,4,4,4,2,3,3,3,3,3,3,3,2,1,1,1,1,1,1,1,2,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,2,1,1,3,3,1,1,1,1,2,3,3,3,3,3,3,2,3,3,3],
-[4,4,4,4,4,4,4,4,4,2,3,3,3,3,3,3,3,2,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,2,1,1,4,4,1,1,1,1,2,5,5,5,5,3,3,2,3,3,3],
-[4,4,4,4,4,4,4,4,4,2,3,3,3,3,3,3,3,2,4,4,4,4,4,4,4,1,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,2,4,4,4,4,4,4,4,4,2,1,1,1,1,3,3,2,3,3,3],
-[4,4,4,4,4,4,4,4,4,2,3,3,3,3,3,3,3,1,4,4,4,4,4,4,4,1,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,2,4,4,4,4,4,4,4,4,2,1,1,1,1,3,3,2,3,3,3],
-[4,4,4,4,4,4,4,4,4,1,3,3,3,3,3,3,3,1,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,2,4,4,4,4,4,4,4,4,2,1,1,1,1,3,3,2,3,3,3],
-[4,4,4,4,4,4,4,4,4,1,3,3,3,3,3,3,3,1,4,4,4,4,4,4,4,5,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,2,4,4,4,4,4,4,4,4,2,3,3,3,3,3,3,2,3,3,3],
-[4,4,4,4,4,4,4,4,4,1,3,3,3,3,3,3,3,4,4,4,4,4,4,4,4,2,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,2,4,4,4,4,4,4,4,4,2,3,3,3,3,3,3,2,3,3,3],
-[4,4,4,4,4,4,4,4,4,4,3,3,3,3,3,3,3,4,4,4,4,4,4,4,4,2,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,2,4,4,4,4,4,4,4,4,2,3,3,3,3,3,3,2,3,3,3],
-[4,4,4,4,4,4,4,4,4,4,3,3,3,3,3,3,3,5,4,4,4,4,4,4,4,2,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,2,4,4,4,4,4,4,4,4,2,3,3,3,3,3,3,2,3,3,3],
-[4,4,4,4,4,4,4,4,4,5,3,3,3,3,3,3,3,2,4,4,4,4,4,4,4,2,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,2,4,4,4,4,4,4,4,4,2,3,3,3,3,3,3,2,3,3,3],
-[4,4,4,4,4,4,4,4,4,2,3,3,3,3,3,3,3,2,4,4,4,4,4,4,4,2,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,2,4,4,4,4,4,4,4,4,2,3,3,3,3,3,3,2,3,3,3],
-[4,4,4,4,4,4,4,4,4,2,3,3,3,3,3,3,3,2,4,4,4,4,4,4,4,2,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,2,4,4,4,4,4,4,4,4,2,3,3,3,3,3,3,2,3,3,3],
-[5,5,5,5,5,5,5,5,5,2,3,3,3,3,3,3,3,2,5,5,5,5,5,5,5,2,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,2,5,5,5,5,5,5,5,5,2,3,3,3,3,3,3,2,3,3,3],
-[1,1,1,1,1,1,1,1,1,2,3,3,3,3,3,3,3,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,3,3,3,3,3,3,2,3,3,3],
-[1,1,1,1,1,1,1,1,1,2,3,3,3,3,3,3,3,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,3,3,3,3,3,3,2,3,3,3],
-[1,1,1,1,1,1,1,1,1,2,3,3,3,3,3,3,3,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,3,3,3,3,3,3,2,3,3,3],
-[4,4,4,4,4,4,4,4,4,2,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,2,3,3,3],
-[4,4,4,4,4,4,4,4,4,2,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,2,3,3,3],
-[4,4,4,4,4,4,4,4,4,2,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,2,3,3,3],
-[4,4,4,4,4,4,4,4,4,2,3,3,3,3,3,3,3,5,5,5,5,3,3,5,5,5,5,5,5,5,5,5,5,5,5,3,3,3,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,3,3,3,3,3,2,3,3,3],
-[4,4,4,4,4,4,4,4,4,2,3,3,3,3,3,3,3,2,1,1,1,3,3,1,1,1,1,2,1,1,1,1,1,1,1,3,3,3,1,1,1,1,1,1,1,1,1,2,1,1,1,1,1,2,3,3,3,3,3,2,3,3,3],
-[4,4,4,4,4,4,4,4,4,2,3,3,3,3,3,3,3,2,1,1,1,3,3,1,1,1,1,2,1,1,1,1,1,1,1,3,3,3,1,1,1,1,1,1,1,1,1,2,1,1,1,1,1,2,3,3,3,3,3,2,3,3,3],
-[4,4,4,4,4,4,4,4,4,2,3,3,3,3,3,3,3,2,1,1,1,3,3,1,1,1,1,2,1,1,1,1,1,1,1,3,3,3,1,1,1,1,1,1,1,1,1,2,1,1,1,1,1,2,3,3,3,3,3,2,3,3,3],
-[4,4,4,4,4,4,4,4,4,2,3,3,3,3,3,3,3,2,3,3,3,3,3,3,3,3,3,2,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,2,3,3,3,3,3,2,3,3,3,3,3,2,3,3,3],
-[4,4,4,4,4,4,4,4,4,2,3,3,3,3,3,3,3,2,3,3,3,3,3,3,3,3,3,2,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,2,3,3,3,3,3,2,3,3,3,3,3,2,3,3,3],
-[4,4,4,4,4,4,4,4,4,2,5,5,3,3,3,5,5,2,3,3,3,3,3,3,3,3,3,2,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,2,3,3,3,3,3,2,3,3,3,3,3,2,3,3,3],
-[4,4,4,4,4,4,4,4,4,1,1,1,4,4,4,1,1,2,3,3,3,3,3,3,3,3,3,2,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,2,3,3,3,3,3,2,3,3,3,3,3,2,3,3,3],
-[4,4,4,4,4,4,4,4,4,1,1,1,4,4,4,1,1,2,3,3,3,3,3,3,3,3,3,2,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,2,3,3,3,3,3,2,3,3,3,3,3,2,3,3,3],
-[4,4,4,4,4,4,4,4,4,1,1,1,4,4,4,1,1,2,3,3,3,3,3,3,3,3,3,2,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,2,3,3,3,3,3,2,3,3,3,3,3,2,3,3,3],
-[4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,2,3,3,3,3,3,3,3,3,3,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,3,3,3,3,3,2,3,3,3,3,3,2,3,3,3],
-[4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,2,3,3,3,3,3,3,3,3,3,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,3,3,3,3,3,2,3,3,3,3,3,2,3,3,3],
-[4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,2,3,3,3,3,3,3,3,3,3,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,3,3,3,3,3,2,3,3,3,3,3,2,3,3,3],
-[4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,2,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,2,3,3,3,3,3,2,3,3,3],
-[4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,4,2,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,3,2,3,3,3,3,3,2,3,3,3],
-[5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,5,2,5,5,5,5,5,2,2,2,2],
-[1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1],
-[1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1],
-[1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1]]
-    
-    def drawMap(self, surface):
-            for i,row in enumerate(self.array):
-                for j,tile in enumerate(row):
-                    if tile > 0:
-                        # if x position is greater than the negative side of the tile and if x position is less than the size of the width of the screen and y position is greater than the size of the tile and y position is less than the height of the screen then render the image of tile
-                        if (self.x + j * self.tilesize) > -self.tilesize and (self.x + j * self.tilesize) < 700 and (self.y + i * self.tilesize) > -self.tilesize and (self.y + i * self.tilesize) < 500:
-                            surface.blit(self.images[tile - 1], (self.x + j * self.tilesize, self.y + i * self.tilesize))
-
-    def move_x(self, direction):
-         self.x += direction
-
-    def move_y(self, direction):
-         self.y += direction
